use int not long

+++ send_signal.c
@@ -25,9 +25,9 @@
 #include <signal.h>
 
 static pid_t check_pid(const char *, const char*);
-void send_signal(char *, long);
+void send_signal(char *, int);
 
-void send_signal(char *executable, long signal_number) {
+void send_signal(char *executable, int signal_number) {
     DIR *processes;
     struct dirent *program;
     pid_t pid;
